+++ vscode/out/gui/CommandRunnerResultView.js
@@ -12,9 +12,9 @@ class CommandRunnerResultView {
         this.script_test_uri = vscode.Uri.file(path.join(this.resource_path, 'crunner_test.js'));
         this.script_rulerunner_result_uri = vscode.Uri.file(path.join(this.resource_path, 'rulerunner_result.js'));
     }
-    get_panel() {
+    get_panel(title) {
         if (this.panel === undefined) {
-            this.panel = vscode.window.createWebviewPanel('commandRunnerResult', 'Command-runner result', vscode.ViewColumn.Beside, { enableScripts: true, retainContextWhenHidden: true });
+            this.panel = vscode.window.createWebviewPanel('commandRunnerResult', title, vscode.ViewColumn.Beside, { enableScripts: true, retainContextWhenHidden: true });
             this.panel.onDidDispose((e) => { this.panel = undefined; });
         }
         if (this.panel !== undefined) {
@@ -25,17 +25,17 @@ class CommandRunnerResultView {
         return this.panel;
     }
     show_rulerunner_result(result) {
-        let panel = this.get_panel();
+        let panel = this.get_panel('Rule-runner result');
         panel.webview.html = this.get_parser_html(this.script_rulerunner_result_uri);
         panel.webview.postMessage({ present_rulerunner_result: result });
     }
     show_test_result(result) {
-        let panel = this.get_panel();
+        let panel = this.get_panel('Command-runner test result');
         panel.webview.html = this.get_parser_html(this.script_test_uri);
         panel.webview.postMessage({ present_test_result: result });
     }
     show_parser_result(result) {
-        let panel = this.get_panel();
+        let panel = this.get_panel('Command-runner parse-only result');
         panel.webview.html = this.get_parser_html(this.script_parser_uri);
         panel.webview.postMessage({ present_parser_result: result });
     }