+++ vscode/src/command-runner/CommandRunner.ts
@@ -14,12 +14,15 @@ export class CommandRunner {
     private commandrunner_uri : vscode.Uri | undefined = undefined;
     private commandrunner_user : string | undefined = undefined;
     private commandrunner_password : string | undefined = undefined;
+    private commandrunner_verbose : boolean | undefined = false;
+    private commandrunner_inject_tags : string[] = [];
 
     public constructor() {
         this.commandrunner_path = vscode.workspace.getConfiguration().get('indeni.commandRunnerPath');
         this.commandrunner_user = vscode.workspace.getConfiguration().get('indeni.commandRunnerUser');
         this.commandrunner_password = vscode.workspace.getConfiguration().get('indeni.commandRunnerPassword');
-
+        this.commandrunner_verbose = vscode.workspace.getConfiguration().get('indeni.commandRunnerVerbose');
+        this.commandrunner_inject_tags = vscode.workspace.getConfiguration().get('indeni.commandRunnerInjectTags') || [];
         if (this.commandrunner_path !== undefined) {
             this.commandrunner_uri = vscode.Uri.file(this.commandrunner_path);
         }
@@ -40,6 +43,39 @@ export class CommandRunner {
         return result;
     }
     
+    private get verbose() : string {
+        if (this.commandrunner_verbose === true) {
+            return "--verbose ";
+        }
+
+        return "";
+    }
+
+    private inject_tags() : string {
+        if (this.commandrunner_inject_tags.length <= 0) {
+            console.log('Inject tags = 0');
+            return '';
+        }
+
+        let result : string[] = [];
+        for (let line of this.commandrunner_inject_tags) {
+            console.log('Parsing line: ' + line);
+            let key_value = line.split('=', 2);
+            console.log(key_value);
+            if (key_value.length === 2) {
+                console.log('Reached here');
+                let item = `""${key_value[0].trim()}""=>""${key_value[1].trim()}""`;
+                result.push(item);
+            }
+        }
+
+        if (result.length > 0) {
+            return '--inject-tags "' + result.join(' ') + '" ';
+        }
+
+        return '';
+    }
+
     public RunFullCommand(input_filename : string, ip_address : string, callback : ((result : CommandRunnerParseOnlyResult) => void)) {
         if (!this.verify_command_runner_path() || this.commandrunner_uri === undefined) {
             return;
@@ -52,7 +88,11 @@ export class CommandRunner {
         if (this.commandrunner_password === undefined) {
             this.commandrunner_password = '';
         }
-        let command = this.escape_filename(this.commandrunner_uri.fsPath) + ` full-command --ssh ${this.commandrunner_user},${this.commandrunner_password} --basic-authentication ${this.commandrunner_user},${this.commandrunner_password} ` + this.escape_filename(input_filename) + " " + ip_address;
+        console.log('Inject tags');
+        console.log(this.inject_tags());
+
+        let command = this.escape_filename(this.commandrunner_uri.fsPath) + ` full-command ${this.verbose}${this.inject_tags()}--ssh ${this.commandrunner_user},${this.commandrunner_password} --basic-authentication ${this.commandrunner_user},${this.commandrunner_password} ` + this.escape_filename(input_filename) + " " + ip_address;
+        console.log('Commandrunner command: ' + command);
         child.exec(command, 
         (error, stdout, stderr) => {
             if (error !== null) {
@@ -75,7 +115,9 @@ export class CommandRunner {
             return;
         }
         
-        child.exec(this.escape_filename(this.commandrunner_uri.fsPath) + " parse-only " + this.escape_filename(filename) + " -f " + this.escape_filename(input_filename), 
+        let exec_string = this.escape_filename(this.commandrunner_uri.fsPath) + " parse-only " + this.verbose + this.escape_filename(filename) + " -f " + this.escape_filename(input_filename);
+        console.log(exec_string);
+        child.exec(exec_string, 
         (error, stdout, stderr) => {
             if (error !== null) {
                 console.error(error);
@@ -99,16 +141,6 @@ export class CommandRunner {
             return Promise.reject('Invalid command-runner path');
         }
         
-        let case_name : string | undefined = undefined;
-        await vscode.window.showInputBox({ placeHolder: 'New test case name' }).then(value => 
-            {
-                case_name = value;
-            });
-
-        if (case_name === undefined) {
-            return Promise.reject('No test case name selected');
-        }
-
         let test_case_map : {[key : string] : string} = {};
         const items : vscode.QuickPickItem[] = [];
         let test_cases = split_script.get_test_cases();
@@ -121,9 +153,28 @@ export class CommandRunner {
                     test_case_map[t_case.name] = t_case.input_data_path;
                 }
             }
-            items.push({ label: 'Browse...' });
         }
 
+        let test_case_name : string = '';
+        items.push({ label: 'New...' });
+        let input_test_case_option = await vscode.window.showQuickPick(items, { placeHolder: 'Select existing test case name or select new name' });
+        if (input_test_case_option === undefined) {
+            return Promise.reject('No test case name specified');
+        }
+
+        if (input_test_case_option.label ===  'New...') {
+            let new_name = await vscode.window.showInputBox({ placeHolder: 'Test case name' });
+            if (new_name === undefined) {
+                return Promise.reject('No test case name specified');
+            }
+            test_case_name = new_name;
+        }
+        else {
+            test_case_name = input_test_case_option.label;
+        }
+
+        items.pop();
+        items.push({ label: 'Browse...' });
         let input_filename = await vscode.window.showQuickPick(items, { placeHolder: 'Select existing input file or browse for new' });
         if (input_filename === undefined) {
             return Promise.reject('No input file specified');
@@ -149,7 +200,7 @@ export class CommandRunner {
 
         let script_filename = split_script.header_section !== undefined? split_script.header_section.filename: '';
         
-        let command = "test create " + this.escape_filename(script_filename) + " " + case_name + " " + this.escape_filename(input_data_path);
+        let command = "test create " + this.escape_filename(script_filename) + " " + test_case_name + " " + this.escape_filename(input_data_path);
         let result = await this.RunCommandRunner(command);