+++ vscode/src/gui/CommandRunnerResultView.ts
@@ -20,9 +20,9 @@ export class CommandRunnerResultView {
         this.script_rulerunner_result_uri = vscode.Uri.file(path.join(this.resource_path, 'rulerunner_result.js'));
     }
 
-    private get_panel() : vscode.WebviewPanel {
+    private get_panel(title : string) : vscode.WebviewPanel {
         if (this.panel === undefined) {
-            this.panel = vscode.window.createWebviewPanel('commandRunnerResult', 'Command-runner result', vscode.ViewColumn.Beside, { enableScripts: true, retainContextWhenHidden: true });
+            this.panel = vscode.window.createWebviewPanel('commandRunnerResult', title, vscode.ViewColumn.Beside, { enableScripts: true, retainContextWhenHidden: true });
             this.panel.onDidDispose((e : void) => { this.panel = undefined; });
         }
 
@@ -37,19 +37,19 @@ export class CommandRunnerResultView {
     }
 
     public show_rulerunner_result(result : RuleRunnerCompileResult) {
-        let panel = this.get_panel();
+        let panel = this.get_panel('Rule-runner result');
         panel.webview.html = this.get_parser_html(this.script_rulerunner_result_uri);
         panel.webview.postMessage({ present_rulerunner_result : result });
     }
 
     public show_test_result(result : CommandRunnerTestRunResult) {
-        let panel = this.get_panel();
+        let panel = this.get_panel('Command-runner test result');
         panel.webview.html = this.get_parser_html(this.script_test_uri);
         panel.webview.postMessage( { present_test_result: result });
     }
 
     public show_parser_result(result : CommandRunnerParseOnlyResult) {
-        let panel = this.get_panel();
+        let panel = this.get_panel('Command-runner parse-only result');
         panel.webview.html = this.get_parser_html(this.script_parser_uri);
         panel.webview.postMessage({ present_parser_result : result });
     }